Add title pluralization
