+++ client/landing/stepper/declarative-flow/internals/steps-repository/domain-transfer-complete/styles.scss
@@ -82,7 +82,7 @@
 					p {
 						font-size: 0.875rem;
 					}
-					button {
+					a.components-button {
 						border-radius: 4px;
 						border: 1px solid var(--studio-gray-10);
 						padding: 10px 24px;
